refactor: move cache blockprocess to LedgeVerifier
