+++ ledger/ledger.go
@@ -529,110 +529,3 @@ func (l *Ledger) GenerateChangeBlock(account types.Address, representative types
 	}
 	return &sb, nil
 }
-
-func (l *Ledger) BlockCacheProcess(block types.Block) error {
-	blk := block.(*types.StateBlock)
-	am, err := l.GetAccountMeta(blk.GetAddress())
-	if err != nil && err != ErrAccountNotFound {
-		return fmt.Errorf("get account meta cache error: %s", err)
-	}
-	return l.BatchUpdate(func(txn db.StoreTxn) error {
-		if state, ok := block.(*types.StateBlock); ok {
-			err := l.processCacheBlock(state, am, txn)
-			if err != nil {
-				l.logger.Error(fmt.Sprintf("%s, cache block:%s", err.Error(), state.GetHash().String()))
-				return err
-			}
-			return nil
-		} else if _, ok := block.(*types.SmartContractBlock); ok {
-			return errors.New("smart contract block")
-		}
-		return errors.New("invalid block")
-	})
-}
-
-func (l *Ledger) processCacheBlock(block *types.StateBlock, am *types.AccountMeta, txn db.StoreTxn) error {
-	l.logger.Debug("process block cache, ", block.GetHash())
-	if err := l.AddBlockCache(block, txn); err != nil {
-		return err
-	}
-	if err := l.updateAccountMetaCache(block, am, txn); err != nil {
-		return fmt.Errorf("update account meta cache error: %s", err)
-	}
-	if err := l.updatePendingKind(block, txn); err != nil {
-		return fmt.Errorf("update pending kind error: %s", err)
-	}
-	return nil
-}
-
-func (l *Ledger) updateAccountMetaCache(block *types.StateBlock, am *types.AccountMeta, txn db.StoreTxn) error {
-	hash := block.GetHash()
-	rep := block.GetRepresentative()
-	address := block.GetAddress()
-	token := block.GetToken()
-	balance := block.GetBalance()
-
-	tmNew := &types.TokenMeta{
-		Type:           token,
-		Header:         hash,
-		Representative: rep,
-		OpenBlock:      hash,
-		Balance:        balance,
-		BlockCount:     1,
-		BelongTo:       address,
-		Modified:       common.TimeNow().UTC().Unix(),
-	}
-
-	if am != nil {
-		tm := am.Token(block.GetToken())
-		if block.GetToken() == common.ChainToken() {
-			am.CoinBalance = balance
-			am.CoinOracle = block.GetOracle()
-			am.CoinNetwork = block.GetNetwork()
-			am.CoinVote = block.GetVote()
-			am.CoinStorage = block.GetStorage()
-		}
-		if tm != nil {
-			tm.Header = hash
-			tm.Representative = rep
-			tm.Balance = balance
-			tm.BlockCount = tm.BlockCount + 1
-			tm.Modified = common.TimeNow().UTC().Unix()
-		} else {
-			am.Tokens = append(am.Tokens, tmNew)
-		}
-		if err := l.AddOrUpdateAccountMetaCache(am, txn); err != nil {
-			return err
-		}
-	} else {
-		account := types.AccountMeta{
-			Address: address,
-			Tokens:  []*types.TokenMeta{tmNew},
-		}
-
-		if block.GetToken() == common.ChainToken() {
-			account.CoinBalance = balance
-			account.CoinOracle = block.GetOracle()
-			account.CoinNetwork = block.GetNetwork()
-			account.CoinVote = block.GetVote()
-			account.CoinStorage = block.GetStorage()
-		}
-		if err := l.AddAccountMetaCache(&account, txn); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (l *Ledger) updatePendingKind(block *types.StateBlock, txn db.StoreTxn) error {
-	if block.IsReceiveBlock() {
-		pendingKey := &types.PendingKey{
-			Address: block.GetAddress(),
-			Hash:    block.GetLink(),
-		}
-		if err := l.UpdatePending(pendingKey, types.PendingUsed, txn); err != nil {
-			return err
-		}
-	}
-	return nil
-}
