+++ rpc/api/ledger.go
@@ -787,7 +787,8 @@ func (l *LedgerApi) Process(block *types.StateBlock) (types.Hash, error) {
 	switch flag {
 	case process.Progress:
 		hash := block.GetHash()
-		err := l.ledger.BlockCacheProcess(block)
+		verify := process.NewLedgerVerifier(l.ledger)
+		err := verify.BlockCacheProcess(block)
 		if err != nil {
 			l.logger.Errorf("Block %s add to blockCache error[%s]", hash, err)
 			return types.ZeroHash, err
